refactor: reorganise
