+++ src/app/screens/Accounts/index.tsx
@@ -13,7 +13,7 @@ import utils from "../../../common/lib/utils";
 import type { Account } from "../../../types";
 import Button from "../../components/Button";
 import Container from "../../components/Container";
-import TextField from "../../components/Form/TextField";
+import TextField from "../../components/form/TextField";
 import Menu from "../../components/Menu";
 import { useAccounts } from "../../context/AccountsContext";
 import { useAuth } from "../../context/AuthContext";
