+++ src/app/screens/Enable/index.tsx
@@ -1,6 +1,6 @@
 import { useState, useEffect, useCallback, useRef } from "react";
 
-import ConfirmOrCancel from "../../components/layout/ConfirmOrCancel";
+import ConfirmOrCancel from "../../components/ConfirmOrCancel";
 import PublisherCard from "../../components/PublisherCard";
 import msg from "../../../common/lib/msg";
 import type { OriginData } from "../../../types";
