+++ src/app/screens/Onboard/ConnectCitadel/index.tsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 
-import Input from "../../../components/Form/Input";
+import Input from "../../../components/form/Input";
 import Button from "../../../components/Button";
 import { useNavigate } from "react-router-dom";
