+++ src/app/screens/Onboard/ConnectLndHub/index.tsx
@@ -5,7 +5,7 @@ import utils from "../../../../common/lib/utils";
 
 import Button from "../../../components/Button";
 import QrcodeScanner from "../../../components/QrcodeScanner";
-import TextField from "../../../components/Form/TextField";
+import TextField from "../../../components/form/TextField";
 import CompanionDownloadInfo from "../../../components/CompanionDownloadInfo";
 
 export default function ConnectLndHub() {
