+++ src/app/screens/Onboard/SetPassword/index.tsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import TextField from "../../../components/Form/TextField";
+import TextField from "../../../components/form/TextField";
 import Button from "../../../components/Button";
 import { useNavigate } from "react-router-dom";
 import utils from "../../../../common/lib/utils";
