+++ src/app/screens/Receive.tsx
@@ -15,7 +15,7 @@ import { useAuth } from "../context/AuthContext";
 
 import Button from "../components/Button";
 import IconButton from "../components/IconButton";
-import TextField from "../components/Form/TextField";
+import TextField from "../components/form/TextField";
 // import Select from "../components/Form/Select";
 import Header from "../components/Header";
 import Loading from "../components/Loading";
